% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_fichas_empresa_universia.R
\name{scrap_fichas_empresas_universia}
\alias{scrap_fichas_empresas_universia}
\title{Scrapea el directorio de empresas de Universia.}
\usage{
scrap_fichas_empresas_universia(municipio, provincias)
}
\arguments{
\item{municipio, }{provincias}
}
\value{
json
}
\description{
Scrapea el directorio de empresas de Universia devolviendo un json resultado y envíandolo a la plataforma Smart City.
}
\examples{
scrap_fichas_empresas_universia('Camargo', 'Cantabria')

}
